+++ src/Phonebook.Core/ViewModels/ContactsViewModel.cs
@@ -11,75 +11,93 @@ namespace Phonebook.Core.ViewModels
 {
     public class ContactsViewModel : MvxViewModel
     {
-        private int page = 1;
-        private MvxObservableCollection<ItemContact> _items;
-        private bool _isRefreshing;
-      
-        readonly IContactService _contactService;
-        readonly IMvxNavigationService _navigationService;
-        readonly IMessage _message;
 
+        #region Fields
+        private int _page = 1;
+        #endregion
+
+        #region Commands
         private IMvxCommand _refreshContactsCommand;
-        private IMvxCommand _gettingCommand;
+        public IMvxCommand RefreshContactsCommand => _refreshContactsCommand ?? (_refreshContactsCommand = new MvxAsyncCommand(PullToRefresh));
 
+        private IMvxCommand _gettingCommand;
+        public IMvxCommand GettingContactsCommand => _gettingCommand ?? (_gettingCommand = new MvxAsyncCommand(GetContacts));
+        #endregion
+
+        #region Properties
+
+        private MvxObservableCollection<ItemContact> _items;
         public MvxObservableCollection<ItemContact> Items
         {
             get { return _items; }
             set { SetProperty(ref _items, value); }
         }
+
+        private bool _isRefreshing;
         public bool IsRefreshing
         {
             get => _isRefreshing;
             set => SetProperty(ref _isRefreshing, value);
         }
+        #endregion
 
-        public IMvxCommand RefreshContactsCommand => _refreshContactsCommand ?? (_refreshContactsCommand = new MvxAsyncCommand(PullToRefresh));
-        public IMvxCommand GettingContactsCommand => _gettingCommand ?? (_gettingCommand = new MvxAsyncCommand(GetContacts));
+        #region Services
+        protected IContactService ContactService { get; }
+        protected IMvxNavigationService NavigationService { get; }
+        protected IMessage Message { get; }
+        #endregion
+
+        #region Constructors
         public ContactsViewModel(IMvxNavigationService navigationService, IContactService contactService, IMessage message)
         {
-            _navigationService = navigationService;
-            _contactService = contactService;
-            _message = message;
+            NavigationService = navigationService;
+            ContactService = contactService;
+            Message = message;
             _items = new MvxObservableCollection<ItemContact>();
         }
+        #endregion
 
+        #region Private
         private async Task GetContacts()
         {
             try
             {
-                var contacts = await _contactService.GetContacts(SettingsConstants.CountOfContacts, page).ConfigureAwait(false);
+                var contacts = await ContactService.GetContacts(SettingsConstants.CountOfContacts, _page).ConfigureAwait(false);
                 if (contacts != null && contacts.Contacts.Count == SettingsConstants.CountOfContacts)
                 {
-                    page++;
+                    _page++;
                     var tmp = new MvxObservableCollection<ItemContact>();
                     foreach (var cont in contacts.Contacts)
                         tmp.Add(new ItemContact(cont));
                     Items.AddRange(tmp);
                 }
                 else
-                   Mvx.IoCProvider.Resolve<IMessage>().Dialog("Error: Connection problem or server is not responding!", async () => { await PullToRefresh(); });
+                    Message.Dialog("Error: Connection problem or server is not responding!", async () => { await PullToRefresh(); });
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
-                Mvx.IoCProvider.Resolve<IMessage>().Dialog($"Error: {ex.Message}", async () => { await PullToRefresh(); });
+                Message.Dialog($"Error: {ex.Message}", async () => { await PullToRefresh(); });
             }
-           
         }
-
         private async Task PullToRefresh()
         {
             IsRefreshing = true;
-            page = 1;
+            _page = 1;
             Items.Clear();
             await GetContacts();
             IsRefreshing = false;
         }
+        #endregion
 
-        public override void ViewAppeared()
+        #region Protected
+        #endregion
+
+        #region Public
+        public override void ViewCreated()
         {
-            if(Items.Count == 0)
-                Task.Run(PullToRefresh);
-            base.ViewAppeared();
+            Task.Run(PullToRefresh);
+            base.ViewCreated();
         }
+        #endregion
     }
 }
