--- src/Phonebook.Core/ViewModels/Details/DetailsViewModel.cs
@@ -1,11 +1,0 @@
-﻿using MvvmCross.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Phonebook.Core.ViewModels.Details
-{
-    class DetailsViewModel : MvxViewModel
-    {
-    }
-}
