--- src/Phonebook.Core/ViewModels/Photo/PhotoViewModel.cs
@@ -1,12 +1,0 @@
-﻿using MvvmCross.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Phonebook.Core.ViewModels.Photo
-{
-    class PhotoViewModel : MvxViewModel
-    {
-
-    }
-}
