+++ src/Phonebook.Droid/Phonebook.Droid.csproj
@@ -60,7 +60,7 @@
   <ItemGroup>
     <Compile Include="LinkerPleaseInclude.cs" />
     <Compile Include="MessageDroid.cs" />
-    <Compile Include="Resources\AdapterContacts.cs" />
+    <Compile Include="AdapterContacts.cs" />
     <Compile Include="Setup.cs" />
     <Compile Include="Views\ContactsView.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
@@ -171,16 +171,6 @@
   </ItemGroup>
   <ItemGroup>
     <AndroidResource Include="Resources\drawable\border.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <AndroidResource Include="Resources\layout\DetailsView.axml">
-      <SubType>Designer</SubType>
-    </AndroidResource>
-  </ItemGroup>
-  <ItemGroup>
-    <AndroidResource Include="Resources\layout\PhotoView.axml">
-      <SubType>Designer</SubType>
-    </AndroidResource>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
