--- src/Phonebook.Droid/Resources/layout/DetailsView.axml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-	<LinearLayout
-	xmlns:android			="http://schemas.android.com/apk/res/android"
-	xmlns:local				="http://schemas.android.com/apk/res-auto"
-	android:orientation		="vertical"
-	android:layout_width	="match_parent"
-	android:layout_height	="match_parent"
-	android:background		="#394d5b"
-	android:paddingLeft="20dp"
-	android:paddingRight="20dp">
-	<ffimageloading.cross.MvxCachedImageView
-		local:MvxBind			="ImagePath Photo"
-		android:layout_width    ="120dp"
-		android:layout_height	="120dp"
-		android:layout_gravity	="center"
-		android:src				="@drawable/abc_ab_share_pack_mtrl_alpha"
-		android:layout_marginTop="20dp" 
-		/>
-	<TextView
-		android:text="FullName"
-		android:textSize="30dp"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:id="@+id/textView1"
-		android:gravity="center" />
-	<TextView
-		android:text="Phone"
-		android:textSize="20dp"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:id="@+id/textView1" />
-	<TextView
-		android:text="88800553535"
-		android:textSize="20dp"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:id="@+id/textView1"
-		android:layout_marginBottom="20dp" />
-	<TextView
-		android:text="Email"
-		android:textSize="20dp"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:id="@+id/textView1" />
-	<TextView
-		android:text="loymax@loymax.ru"
-		android:textSize="20dp"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:id="@+id/textView1" />
-</LinearLayout>
