+++ src/Phonebook.iOS/Phonebook.iOS.csproj
@@ -158,8 +158,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Resources\" />
-    <Folder Include="Views\Details\" />
-    <Folder Include="Views\Photo\" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
